+++ moirai-iter/src/parallel.rs
@@ -20,9 +20,10 @@ mod tests;
 mod traits;
 
 pub use adapters::{
-    Chain, Chunks, Cloned, Copied, Enumerate, Filter, FilterMap, FlatMap, Flatten, Inspect,
-    Interleave, InterleaveShortest, Intersperse, Map, MapInit, MapPositions, MapWith, PanicFuse,
-    Positions, Rev, Skip, SkipAnyWhile, StepBy, Take, TakeAnyWhile, Update, WhileSome, Zip, ZipEq,
+    Chain, Chunks, Cloned, Copied, Enumerate, ExponentialBlocks, Filter, FilterMap, FlatMap,
+    Flatten, Inspect, Interleave, InterleaveShortest, Intersperse, Map, MapInit, MapPositions,
+    MapWith, PanicFuse, Positions, Rev, Skip, SkipAnyWhile, StepBy, Take, TakeAnyWhile,
+    UniformBlocks, Update, WhileSome, Zip, ZipEq,
 };
 pub use consumers::Reduction;
 pub use fallible::TryStreamItem;